+++ src/main/java/com/xiliulou/electricity/mapper/WechatWithdrawalCertificateMapper.java
@@ -30,7 +30,7 @@ public interface WechatWithdrawalCertificateMapper extends BaseMapper<WechatWith
      * @author caobotao.cbt
      * @date 2024/6/12 13:51
      */
-    int logicalDeleteByPayParamsId(@Param("payParamsId") Long payParamsId, @Param("tenantId") Integer tenantId);
+    int logicalDeleteByPayParamsId(@Param("payParamsId") Integer payParamsId, @Param("tenantId") Integer tenantId);
     
     /**
      * 根据租户+加盟商id查询
